macOS: move services menu support into SurfaceView+Services.swift
Pure code motion out of Ghostty.SurfaceView.swift; the only change is
dropping the MARK header the file name now carries. The documentation link
moves with the extension.

NSTextInputClient stays behind: it touches four private members
(markedText, keyTextAccumulator, leadSurrogate, committedPreeditTextAction)
shared with the key event path in the class body, so moving it would mean
widening those to internal.
